Using an LRU cache instead of a beaker cache for the file process. Fixed a transcoder bug if a file fails in the fileprocess after transcoding. Also dropped some index that was causing problems in the sqlite db.
--HG--
branch : trunk

+++ fileprocess/fileprocess/actions/transcoder.py
@@ -5,6 +5,7 @@ from fileprocess.configuration import config
 from fileprocess.processingthread import na
 from baseaction import BaseAction
 import guid
+from hashlib import sha1
 
 log = logging.getLogger(__name__)
 
@@ -45,7 +46,8 @@ class Transcoder(BaseAction):
                 self.cleanup(file)
                 return False
 
-            new_fname = path.join(config['media_dir'], guid.generate())
+            new_fname = \
+                path.join(config['upload_dir'], file['fbid'], guid.generate())
 
             res = os.system(
                'nice -n +17 faad -q -o - %s 2>/dev/null | nice -n +17 lame --quiet -b 128 - %s' %
@@ -54,12 +56,24 @@ class Transcoder(BaseAction):
                 log.info('Error transcoding mp4 file')
                 file['msg'] = 'Transcoding error'
                 file['na'] = na.FAILURE
-                os.remove(new_fname)
+                if os.path.exists(new_fname):
+                    os.remove(new_fname)
                 self.cleanup(file)
                 return False
             
             os.remove(file['fname'])
             file['fname'] = new_fname
+
+            new_sha = sha1(open(new_fname, 'rb').read()).hexdigest()
+            new_fname = path.join(config['upload_dir'], file['fbid'], new_sha)
+            try:
+                os.rename(file['fname'], new_fname)
+            except OSError:
+                self.cleanup(file)
+                return False
+
+            file['fname'] = new_fname
+            file['usersha'] = new_sha
 
             return file
         else:
